refactor(pythonista): extract MergerUI mixins for T004 complexity
Split the largest MergerUI method groups out of build.py into explicit mixins
while preserving the dual package/flat import contract and pythonista behavior.
This is a follow-up slice of REPOGROUND-LEGACY-RECONCILIATION-V1-T004 and does
not claim full task closeout.
