+++ merger/repoground/frontends/pythonista/build.py
@@ -72,6 +72,9 @@ if __package__:
         resolve_pre_pull_switch_value,
         run_pre_pull_two_phase as _run_pre_pull_two_phase,
     )
+    from merger.repoground.frontends.pythonista.merger_ui_prescan import MergerUIPrescanMixin
+    from merger.repoground.frontends.pythonista.merger_ui_merge_run import MergerUIMergeRunMixin
+    from merger.repoground.frontends.pythonista.merger_ui_browser import MergerUIBrowserMixin
 else:
     from build_utils import normalize_path, normalize_repo_id, safe_script_path
     from build_helpers import deserialize_prescan_pool, resolve_pool_include_paths
@@ -93,6 +96,9 @@ else:
         resolve_pre_pull_switch_value,
         run_pre_pull_two_phase as _run_pre_pull_two_phase,
     )
+    from merger_ui_prescan import MergerUIPrescanMixin
+    from merger_ui_merge_run import MergerUIMergeRunMixin
+    from merger_ui_browser import MergerUIBrowserMixin
 
 
 # Compatibility surface consumed by tests and the dependency-injected CLI runner.
@@ -654,7 +660,7 @@ def run_ui(hub: Path) -> int:
     v.present('fullscreen', hide_title_bar=True)
     return 0
 
-class MergerUI(object):
+class MergerUI(MergerUIPrescanMixin, MergerUIBrowserMixin, MergerUIMergeRunMixin):
     def __init__(self, hub: Path) -> None:
         self.hub = hub
         self.repos = find_repos_in_hub(hub)
@@ -1770,1816 +1776,57 @@ class MergerUI(object):
         except Exception:
             return 0
 
-    def merge_pr_schau_bundles(self, ds, items, sheet) -> None:
-        selected_indices = ds.selected
-        if not selected_indices:
-            if console:
-                console.hud_alert("No bundles selected", "error")
-            return
-
-        # Deterministische Auswahlreihenfolge: sortiere Indices nach Timestamp (desc) der Items
-        # Dies ist robuster als das Sortieren der extrahierten Liste
-        # Note: We rely on string sorting of 'ts' which is strictly ISO-8601-like (%Y-%m-%dT%H%M%SZ)
-        sorted_indices = sorted(selected_indices, key=lambda i: items[i]["ts"], reverse=True)
-        selected_items = [items[i] for i in sorted_indices]
-
-        # Zielverzeichnis: merges/
-        merges_dir = get_merges_dir(self.hub)
-        now_ts = datetime.datetime.now(datetime.timezone.utc).strftime("%Y-%m-%dT%H%M%SZ")
-        out_filename = f"pr-schau-merge_{now_ts}.md"
-        out_path = merges_dir / out_filename
-
-        lines = [
-            "# PR-Schau Merge Report",
-            f"- Generated: {now_ts}",
-            f"- Source root: {PR_SCHAU_DIR}",
-            f"- Bundles: {len(selected_items)}",
-            "",
-            "## Included Bundles",
-        ]
-
-        # Inhaltsverzeichnis
-        for item in selected_items:
-            lines.append(f"- {item['display']}")
-        lines.append("")
-
-        MAX_CHARS = 40000
-
-        for item in selected_items:
-            bdir = item.get("bundle_dir")
-            if not bdir: continue
-
-            # Load metadata
-            meta = {}
-            delta = {}
-            try:
-                meta_raw, _ = load_pr_schau_bundle(bdir, strict=False, verify_level="none")
-                meta = _flatten_meta(meta_raw)
-            except Exception:
-                pass
-
-            try:
-                p_delta = bdir / "delta.json"
-                if p_delta.exists():
-                    delta = json.loads(p_delta.read_text("utf-8"))
-            except Exception:
-                pass
-
-            repo = meta.get("repo", item["repo"])
-            created = meta.get("created_at", item["ts"])
-
-            summary_str = "n/a"
-            if "summary" in delta:
-                s = delta["summary"]
-                summary_str = f"+{s.get('added',0)} / ~{s.get('changed',0)} / -{s.get('removed',0)}"
-
-            lines.append(f"## {repo} @ {created}")
-            # Provenance: Bundle-Pfad relativ zum Hub (falls möglich) oder zu Source Root
-            try:
-                rel_bundle_path = bdir.relative_to(self.hub)
-                lines.append(f"- Bundle dir: `{rel_bundle_path}`")
-            except Exception:
-                lines.append(f"- Bundle dir: `{bdir.name}`")
-
-            lines.append(f"- **Summary**: {summary_str}")
-            if "hub_rel" in meta:
-                lines.append(f"- **Path**: `{meta['hub_rel']}`")
-            if "old_tree_hint" in meta:
-                lines.append(f"- **Base**: `{meta['old_tree_hint']}`")
-            if "new_tree_hint" in meta:
-                lines.append(f"- **Head**: `{meta['new_tree_hint']}`")
-            lines.append("")
-
-            # Review Content (Truncated)
-            review_md = item.get("path")
-            if review_md and review_md.exists():
-                lines.append("### Review Content")
-                try:
-                    content = review_md.read_text("utf-8", errors="replace")
-                    if len(content) > MAX_CHARS:
-                        lines.append(f"> **Note**: Content truncated at {MAX_CHARS} characters. Full content in original `review.md`.")
-                        content = content[:MAX_CHARS] + "\n\n... [Truncated due to size] ..."
-                    lines.append(content)
-                except Exception as e:
-                    lines.append(f"> Error reading review content: {e}")
-            else:
-                lines.append("> (No review.md content available)")
-            lines.append("")
-            lines.append("---")
-            lines.append("")
-
-        try:
-            out_path.write_text("\n".join(lines), encoding="utf-8")
-
-            # Validierung vor Feedback
-            if not out_path.exists() or out_path.stat().st_size == 0:
-                raise RuntimeError("Output file empty or missing.")
-
-            msg = f"Merged {len(selected_items)} bundles to {out_filename}"
-            if console:
-                console.hud_alert("Merge created", "success")
-            else:
-                print(msg)
-
-            # Optional: Open output if editor available
-            if editor:
-                editor.open_file(str(out_path))
-
-            # Close sheet on success
-            sheet.close()
-
-        except Exception as e:
-            if console:
-                console.alert("Merge Failed", str(e), "OK")
-            else:
-                print(f"Merge Failed: {e}")
-
-    def show_pr_schau_browser(self, sender):
-        """Zeigt Liste der verfügbaren PR-Schau Bundles mit Multi-Select Workflow."""
-        pr_dir = self.hub / PR_SCHAU_DIR
-
-        def _normalize_ts(val: str) -> Optional[str]:
-            """Ensure timestamp is strictly %Y-%m-%dT%H%M%SZ."""
-            # 0. Safety guard for non-string types (e.g. from JSON)
-            if not isinstance(val, str):
-                return None
-            # 1. Check strict match
-            if re.match(r"^\d{4}-\d{2}-\d{2}T\d{6}Z$", val):
-                return val
-            # 2. Try parsing common formats
-            try:
-                # Handle Z suffix manually for pre-3.11 compatibility or partial ISO
-                clean = val.replace("Z", "+00:00")
-                dt = datetime.datetime.fromisoformat(clean)
-                # Re-format strict
-                return dt.strftime("%Y-%m-%dT%H%M%SZ")
-            except ValueError:
-                return None
-
-        items = []
-        if pr_dir.exists():
-            for repo_dir in pr_dir.iterdir():
-                if not repo_dir.is_dir(): continue
-                repo_name = repo_dir.name
-
-                for ts_dir in repo_dir.iterdir():
-                    if not ts_dir.is_dir(): continue
-
-                    # Timestamp Contract: Ensure strictly formatted timestamp folder name
-                    # Expected: %Y-%m-%dT%H%M%SZ (e.g. 2025-05-10T123000Z)
-                    ts_raw = ts_dir.name
-                    ts_sort = _normalize_ts(ts_raw)
-
-                    if not ts_sort:
-                        # Attempt fallback from bundle metadata if folder name is invalid
-                        try:
-                            # Use canonical loader even for fallback checks
-                            bj_raw, _ = load_pr_schau_bundle(ts_dir, strict=False, verify_level="none")
-                            bj = _flatten_meta(bj_raw)
-
-                            # legacy created_at / v1 generated_at / fallback ts
-                            candidates = [bj.get("created_at"), bj.get("generated_at"), bj.get("ts")]
-                            ts_source = next((x for x in candidates if x), None)
-                            if ts_source:
-                                # Normalize the JSON timestamp too
-                                ts_sort = _normalize_ts(ts_source)
-                        except Exception:
-                            pass
-
-                    # If still no valid sort key, use a fallback to ensure list display but minimal priority
-                    display_ts = ts_raw
-                    if not ts_sort:
-                        ts_sort = "0000-00-00T000000Z" # Sorts to bottom in desc
-                        display_ts = f"{ts_raw} (invalid ts)"
-                    else:
-                        display_ts = ts_sort
-
-                    review_md_path = ts_dir / "review.md"
-                    bundle_json_path = ts_dir / BUNDLE_FILENAME
-                    delta_json_path = ts_dir / "delta.json"
-
-                    # Robustness: Include even if review.md missing, if metadata exists
-                    if review_md_path.exists() or bundle_json_path.exists() or delta_json_path.exists():
-                        display_text = f"{repo_name} @ {display_ts}"
-                        if not review_md_path.exists():
-                            display_text += " (no review.md)"
-
-                        items.append({
-                            "repo": repo_name,
-                            "ts": ts_sort,
-                            "path": review_md_path,
-                            "bundle_dir": ts_dir,
-                            "display": display_text
-                        })
-
-        if not items:
-            if console:
-                console.alert("PR-Schau", "Keine PR-Bundles gefunden.", "OK", hide_cancel_button=True)
-            return
-
-        # Sort by timestamp descending
-        items.sort(key=lambda x: x["ts"], reverse=True)
-
-        sheet = ui.View()
-        sheet.name = "PR-Schau Bundles"
-        sheet.background_color = "#111111"
-        # Increase size for better overview
-        sheet.frame = (0, 0, 600, 700)
-
-        # Button Bar Area
-        bar_height = 50
-
-        tv = ui.TableView()
-        tv.frame = (0, 0, sheet.width, sheet.height - bar_height)
-        tv.flex = "WH"
-        tv.background_color = "#111111"
-        tv.separator_color = "#333333"
-        tv.row_height = 44 # Better touch target
-
-        ds = PRSchauDataSource(items)
-        tv.data_source = ds
-        tv.delegate = ds
-
-        sheet.add_subview(tv)
-
-        # Bottom Bar
-        bar = ui.View()
-        bar.frame = (0, sheet.height - bar_height, sheet.width, bar_height)
-        bar.flex = "WT"
-        bar.background_color = "#222222"
-        sheet.add_subview(bar)
-
-        btn_y = 8
-        btn_h = 34
-        margin = 10
-
-        # Button: Open (Left aligned, Fixed)
-        btn_open = ui.Button(title="Open")
-        btn_open.frame = (margin, btn_y, 80, btn_h)
-        btn_open.flex = ""
-        btn_open.background_color = "#333333"
-        btn_open.tint_color = "white"
-        btn_open.corner_radius = 6
-
-        def action_open(sender):
-            row = -1
-            # Prioritize last tapped, then first selected
-            if ds.last_tapped_row >= 0:
-                row = ds.last_tapped_row
-            elif ds.selected:
-                row = next(iter(ds.selected))
-
-            if row < 0 or row >= len(items):
-                _notify("Select a bundle to open", "info")
-                return
-
-            # Smart Open: Try review.md -> bundle metadata -> delta.json
-            item = items[row]
-            candidates = [
-                item.get("path"),                   # review.md
-                item.get("bundle_dir") / BUNDLE_FILENAME,
-                item.get("bundle_dir") / "delta.json"
-            ]
-
-            opened = False
-            for cand in candidates:
-                if cand and isinstance(cand, Path) and cand.exists():
-                    path_str = str(cand)
-
-                    # Strategy 1: Editor
-                    if editor:
-                        try:
-                            editor.open_file(path_str)
-                            opened = True
-                            break
-                        except Exception:
-                            pass
-
-                    # Strategy 2: Console Quicklook
-                    if console:
-                        try:
-                            console.quicklook(path_str)
-                            opened = True
-                            break
-                        except Exception:
-                            pass
-
-                    # Strategy 3: Standard Quicklook module
-                    if quicklook:
-                        try:
-                            quicklook.quicklook(path_str)
-                            opened = True
-                            break
-                        except Exception:
-                            pass
-
-                    # Strategy 4: Fallback UI Alert (inform user file exists but can't be viewed)
-                    if ui:
-                        try:
-                            ui.alert("File exists", f"Cannot open: {cand.name}\n(No viewer available)", "OK", hide_cancel_button=True)
-                            opened = True # Handled in UI
-                            break
-                        except Exception:
-                            pass
-
-            if not opened:
-                _notify("No viewable file found", "error")
-
-        btn_open.action = action_open
-        bar.add_subview(btn_open)
-
-        # Button: Close (Right aligned)
-        btn_close = ui.Button(title="Close")
-        btn_close.frame = (sheet.width - 80 - margin, btn_y, 80, btn_h)
-        btn_close.flex = "L"
-        btn_close.background_color = "#333333"
-        btn_close.tint_color = "white"
-        btn_close.corner_radius = 6
-        btn_close.action = lambda s: sheet.close()
-        bar.add_subview(btn_close)
-
-        # Button: Merge Selected (Middle, Flexible width)
-        # Calculate remaining space
-        mid_x = 80 + margin * 2
-        mid_w = sheet.width - (80 + margin * 2) * 2
-
-        btn_merge = ui.Button(title="Merge selected")
-        btn_merge.frame = (mid_x, btn_y, mid_w, btn_h)
-        btn_merge.flex = "W"
-        btn_merge.background_color = "#8E44AD"
-        btn_merge.tint_color = "white"
-        btn_merge.corner_radius = 6
-        btn_merge.action = lambda s: self.merge_pr_schau_bundles(ds, items, sheet)
-
-        bar.add_subview(btn_merge)
-
-        sheet.present("sheet")
-
-    def run_delta_from_last_import(self, sender) -> None:
-        """
-        Erzeugt einen Delta-Merge aus dem neuesten Import-Diff im merges-Ordner.
-        Nutzt die Delta-Helfer aus repolens-extractor.py (falls verfügbar).
-        """
-        merges_dir = get_merges_dir(self.hub)
-        try:
-            candidates = list(merges_dir.glob("*-import-diff-*.md"))
-        except Exception as exc:
-            print(f"[RepoGround] could not scan merges dir: {exc}")
-            candidates = []
-
-        if not candidates:
-            if console:
-                console.alert(
-                    "RepoGround",
-                    "No import diff found.",
-                    "OK",
-                    hide_cancel_button=True,
-                )
-            else:
-                print("[RepoGround] No import diff found.")
-            return
-
-        # jüngstes Diff wählen
-        try:
-            diff_path = max(candidates, key=lambda p: p.stat().st_mtime)
-        except Exception as exc:
-            if console:
-                console.alert(
-                    "RepoGround",
-                    f"Failed to select latest diff: {exc}",
-                    "OK",
-                    hide_cancel_button=True,
-                )
-            else:
-                print(f"[RepoGround] Failed to select latest diff: {exc}")
-            return
-
-        name = diff_path.name
-        prefix = "-import-diff-"
-        if prefix in name:
-            repo_name = name.split(prefix, 1)[0]
-        else:
-            repo_name = name
-
-        repo_root = self.hub / repo_name
-        if not repo_root.exists():
-            msg = f"Repo root not found for diff {diff_path.name}"
-            if console:
-                console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-            else:
-                print(f"[RepoGround] {msg}")
-            return
-
-        mod = _load_repoground_extractor_module()
-        if mod is None or not hasattr(mod, "create_delta_merge_from_diff"):
-            msg = "Delta helper (repolens-extractor) not available."
-            if console:
-                console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-            else:
-                print(f"[RepoGround] {msg}")
-            return
-
-        # Execute delta extraction (without generating a legacy report)
-        try:
-            # We bypass create_delta_merge_from_diff to avoid double-writing.
-            # Instead we extract metadata directly from the diff file.
-            delta_meta = None
-            extract_returned_none = False
-            diff_mtime = None
-
-            try:
-                diff_mtime = diff_path.stat().st_mtime
-            except Exception:
-                diff_mtime = None
-
-            if hasattr(mod, "extract_delta_meta_from_diff_file"):
-                try:
-                    delta_meta = mod.extract_delta_meta_from_diff_file(diff_path)
-                    if delta_meta is None:
-                        extract_returned_none = True
-                except Exception as e:
-                    sys.stderr.write(f"[ERROR] Delta extraction failed: {e}\n")
-
-            # Wenn der Extraktor erfolgreich lief, aber kein Delta liefert, kein Fallback zulassen
-            if extract_returned_none:
-                msg = "Diff enthält keine Delta-Zeilen (keine Änderungen?) – breche ohne Fallback ab."
-                if console:
-                    console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-                else:
-                    print(f"[RepoGround] {msg}")
-                return
-
-            # Fallback nur, wenn der Extraktor nicht verfügbar war oder fehlgeschlagen ist
-            if delta_meta is None:
-                try:
-                    candidate_paths = []
-                    delta_from_diff = diff_path.with_suffix(".delta.json")
-                    candidate_paths.append(delta_from_diff)
-                    try:
-                        repo_specific = sorted(
-                            merges_dir.glob(f"{repo_name}-import-diff-*.delta.json"),
-                            key=lambda p: p.stat().st_mtime,
-                            reverse=True,
-                        )
-                        candidate_paths.extend(repo_specific)
-                    except Exception:
-                        pass
-                    legacy_delta = merges_dir / "delta.json"
-                    candidate_paths.append(legacy_delta)
-
-                    for candidate in candidate_paths:
-                        if not candidate.exists():
-                            continue
-                        # Verhindere veraltete Artefakte: nur Dateien akzeptieren, die zeitlich zum Diff passen
-                        try:
-                            cand_mtime = candidate.stat().st_mtime
-                            if diff_mtime is not None and cand_mtime + 1 < diff_mtime:
-                                continue
-                        except Exception:
-                            pass
-                        raw = json.loads(candidate.read_text(encoding="utf-8"))
-                        if (
-                            isinstance(raw, dict)
-                            and raw.get("type") == "repolens-delta"
-                            and "summary" in raw
-                        ):
-                            delta_meta = raw
-                            break
-                except Exception as e:
-                    print(f"[RepoGround] Failed to read delta metadata: {e}", file=sys.stderr)
-
-            if not delta_meta:
-                msg = "Could not extract delta metadata from diff."
-                if console:
-                    console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-                else:
-                    print(f"[RepoGround] {msg}")
-                return
-
-            # Determine extras config consistent with UI
-            # We use self.extras_config but enable delta_reports
-            extras = ExtrasConfig(
-                health=self.extras_config.health,
-                organism_index=self.extras_config.organism_index,
-                fleet_panorama=self.extras_config.fleet_panorama,
-                augment_sidecar=self.extras_config.augment_sidecar,
-                heatmap=self.extras_config.heatmap,
-                delta_reports=True # Force enable
-            )
-
-            # Need to scan repo for write_reports_v2
-            # Delta Report Strategy:
-            # 1. Scan repo fully (max_bytes=0 => unlimited)
-            # 2. Filter file list based on delta_meta (changed + added)
-            # 3. Use profile 'max' to ensure full content is included for these files
-
-            summary = scan_repo(repo_root, extensions=None, path_contains=None, max_bytes=0, calculate_md5=True, include_hidden=True)
-
-            # Filter files to include only changed/added
-            # Helper to collect paths from delta_meta
-            allowed_paths = set()
-
-            # Check for legacy arrays
-            if "files_added" in delta_meta and isinstance(delta_meta["files_added"], list):
-                allowed_paths.update(delta_meta["files_added"])
-
-            if "files_changed" in delta_meta and isinstance(delta_meta["files_changed"], list):
-                for item in delta_meta["files_changed"]:
-                    if isinstance(item, dict):
-                        path = item.get("path")
-                        if path: allowed_paths.add(path)
-                    elif isinstance(item, str):
-                        allowed_paths.add(item)
-
-            # Check for summary object (repolens-delta schema) if arrays are missing/empty
-            # Note: The schema stores lists in top-level usually. If they are missing, we can't filter.
-            # If allowed_paths is empty but summary says there are changes, we might have a problem.
-            # But assume delta_meta is well-formed from extractor.
-
-            if allowed_paths:
-                # Filter the file list in summary
-                # Use normalized posix string for comparison
-                filtered_files = []
-                for f in summary["files"]:
-                    if f.rel_path.as_posix() in allowed_paths:
-                        filtered_files.append(f)
-
-                summary["files"] = filtered_files
-                # Update stats
-                summary["total_files"] = len(filtered_files)
-                summary["total_bytes"] = sum(f.size for f in filtered_files)
-
-            # Generate merge reports
-            # Use 'max' profile to ensure full content for the changed/added files
-            # (dev/doc logic might otherwise hide content for doc changes)
-            artifacts = write_reports_v2(
-                merges_dir,
-                self.hub,
-                [summary],
-                "max",
-                "repo",
-                0,
-                plan_only=False,
-                code_only=False,
-                debug=False,
-                path_filter=None,
-                ext_filter=None,
-                extras=extras,
-                delta_meta=delta_meta,    # << NEW: real delta injected
-                generator_info={"name": "repoground", "platform": "ios"},
-            )
-
-            # Close files
-            out_paths = artifacts.get_all_paths()
-            force_close_files(out_paths)
-
-            primary_path = artifacts.get_primary_path()
-            msg = (
-                f"Delta report generated: {primary_path.name}"
-                if primary_path is not None
-                else "Delta report generated"
-            )
-            if console:
-                try:
-                    console.hud_alert(msg)
-                except Exception:
-                    console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-            else:
-                print(f"[RepoGround] {msg}")
-
-        except Exception as exc:
-            msg = f"Delta merge failed: {exc}"
-            if console:
-                console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-            else:
-                print(f"[RepoGround] {msg}")
-            return
-
-    def show_prescan_sheet(self, sender):
-        """
-        Shows the Prescan UI (Tree View) for the selected repository.
-        Currently limited to single repo selection for simplicity.
-        
-        ARCHITECTURE:
-        - Prescan → Selection Pool (modify only, never triggers merge)
-        - Merge → Explicit action from main view via Run Merge button
-        - No implicit transition from prescan to merge execution
-        """
-        selected = self._get_selected_repos()
-        if len(selected) != 1:
-            if console:
-                console.hud_alert("Please select exactly one repo for Prescan.", "error")
-            return
-
-        repo_name = selected[0]
-        repo_path = self.hub / repo_name
-
-        # We need to run prescan logic. Since we are in Pythonista (local), we call core directly.
-        try:
-            from merger.repoground.core.merge import prescan_repo
-        except ImportError:
-            try:
-                from merger.repoground.core.merge import prescan_repo
-            except ImportError:
-                _notify("Core merge module not found", "error")
-                return
-
-        # Engage Guard
-        self._prescan_active = True
-
-        # Show Loading HUD
-        if console:
-            console.show_activity("Scanning structure...")
-
-        def run_scan_bg():
-            try:
-                # Run prescan
-                data = prescan_repo(repo_path, max_depth=10)
-                # Ensure UI update on main thread
-                ui.delay(lambda: self._present_prescan_ui(data), 0)
-            except Exception as e:
-                # Bind exception text before scheduling delayed UI callback;
-                # Python clears exception variables after except blocks.
-                _err_msg = str(e)
-                def err():
-                    if console: console.alert("Prescan Failed", _err_msg, "OK", hide_cancel_button=True)
-                    # Reset flag on failure
-                    self._prescan_active = False
-                ui.delay(err, 0)
-            finally:
-                if console:
-                    console.hide_activity()
-
-        # Run in background
-        import threading
-        t = threading.Thread(target=run_scan_bg)
-        t.start()
-
-    def _present_prescan_ui(self, prescan_data):
-        """
-        Displays the prescan tree in a Sheet.
-        Allows selection of files/folders.
-        """
-        root_node = prescan_data["tree"]
-        root_name = prescan_data["root"]
-
-        # Flatten tree for table view (simple indentation approach)
-        flat_items = []
-
-        def traverse(node, depth):
-            # Item struct: { path, display, type, depth, node_ref }
-            name = node["path"].split("/")[-1]
-            if node["path"] == ".": name = root_name
-
-            icon = "📁" if node["type"] == "dir" else "📄"
-            indent = "  " * depth
-            display = f"{indent}{icon} {name}"
-
-            flat_items.append({
-                "path": node["path"],
-                "display": display,
-                "type": node["type"],
-                "depth": depth,
-                "orig": node,
-                "selected": False # Default state
-            })
-
-            if node.get("children"):
-                # Sort: Dirs first, then files
-                dirs = [c for c in node["children"] if c["type"] == "dir"]
-                files = [c for c in node["children"] if c["type"] == "file"]
-
-                dirs.sort(key=lambda x: x["path"])
-                files.sort(key=lambda x: x["path"])
-
-                for c in dirs + files:
-                    traverse(c, depth + 1)
-
-        traverse(root_node, 0)
-
-        # Initially select based on Recommended heuristic
-        # Start with None, then run heuristic
-        for item in flat_items:
-            item["selected"] = False
-
-        # Load existing selection from pool if available
-        # This supports the "Append" workflow by initializing with previous state
-        existing_pool_entry = self.saved_prescan_selections.get(root_name)
-
-        # Logic:
-        # - If pool has entry (dict):
-        #   - If raw is None: ALL state - select everything
-        #   - If raw is list: Use raw for UI truth (not compressed)
-        # - If no pool entry:
-        #   - Run Heuristic (Recommended).
-
-        if existing_pool_entry:
-            if isinstance(existing_pool_entry, dict):
-                raw = existing_pool_entry.get("raw")
-                if raw is None:
-                    # ALL state - select everything
-                    for item in flat_items:
-                        item["selected"] = True
-                elif isinstance(raw, list):
-                    # Partial selection from pool - use raw for UI truth
-                    # Normalize paths for consistent matching
-                    pool_set = set(normalize_path(p) for p in raw)
-                    for item in flat_items:
-                        normalized_item_path = normalize_path(item["path"])
-                        # Direct match
-                        if normalized_item_path in pool_set:
-                            item["selected"] = True
-                        else:
-                            # Check if parent dir is in pool (for compressed paths)
-                            parts = normalized_item_path.split('/')
-                            for i in range(len(parts)):
-                                sub = "/".join(parts[:i+1])
-                                if sub in pool_set:
-                                    item["selected"] = True
-                                    break
-            else:
-                # Legacy format - shouldn't happen after migration
-                # Run heuristic as fallback
-                pass
-        else:
-            # No existing selection -> Heuristic
-            # Run heuristic logic (same as prescanRecommended)
-            def is_recommended(path_str):
-                path = path_str.lower()
-                # Critical
-                if "readme" in path or path.endswith(".ai-context.yml"):
-                    return True
-                # Code
-                parts = path.split('/')
-                if "src" in parts or "contracts" in parts or "docs" in parts:
-                    if "test" not in path:
-                         return True
-                return False
-
-            for item in flat_items:
-                if item["type"] == "file":
-                    if is_recommended(item["path"]):
-                        item["selected"] = True
-
-        # Create Sheet with reliable close handling
-        # ARCHITECTURE NOTE: Prescan → Selection Pool (modify only)
-        # Merge → Explicit action from main view (never triggered from prescan)
-        class PrescanSheet(ui.View):
-            """
-            Custom View subclass.
-            Note: We avoid relying solely on will_close() for critical state reset due to
-            potential delegate limitations/bugs in some Pythonista versions.
-            State is reset explicitly in action handlers.
-            """
-            def __init__(self, parent):
-                super().__init__()
-                self._parent = parent
-                self.name = f"Prescan: {root_name}"
-                self.background_color = "#111111"
-                self.frame = (0, 0, 600, 800)
-
-            def will_close(self):
-                # Fallback safety net
-                if self._parent._prescan_active:
-                     self._parent._prescan_active = False
-
-        sheet = PrescanSheet(self)
-
-        def reset_guard():
-            self._prescan_active = False
-
-        # Track selection mode explicitly for better state management
-        # This helps prevent crashes when transitioning between ALL/PARTIAL/NONE states
-        selection_state = {
-            'mode': 'partial'  # 'all', 'partial', or 'none'
-        }
-        
-        # Initialize selection mode based on current selection
-        # Check if existing pool entry is in ALL state (both raw and compressed are None)
-        is_all = (isinstance(existing_pool_entry, dict) and 
-                  existing_pool_entry.get("raw") is None and 
-                  existing_pool_entry.get("compressed") is None)
-        
-        if is_all:
-            selection_state['mode'] = 'all'
-        elif not any(item["selected"] for item in flat_items):
-            selection_state['mode'] = 'none'
-        else:
-            selection_state['mode'] = 'partial'
-
-        # Header Stats
-        stats_lbl = ui.Label(frame=(10, 10, 580, 20))
-        stats_lbl.text = f"{prescan_data['file_count']} files • {parse_human_size(str(prescan_data['total_bytes']))} bytes total" # approximate
-        stats_lbl.text_color = "gray"
-        stats_lbl.font = ("<System>", 12)
-        sheet.add_subview(stats_lbl)
-
-        # Buttons: Select All / None / Recommended
-        btn_y = 40
-        btn_w = 80
-        btn_h = 30
-
-        def toggle_all(val):
-            for i in flat_items: i["selected"] = val
-            # Update selection mode
-            if val:
-                selection_state['mode'] = 'all'
-            else:
-                selection_state['mode'] = 'none'
-            tv.reload_data()
-
-        btn_all = ui.Button(title="All")
-        btn_all.frame = (10, btn_y, btn_w, btn_h)
-        btn_all.background_color = "#333333"
-        btn_all.tint_color = "white"
-        btn_all.corner_radius = 4
-        btn_all.action = lambda s: toggle_all(True)
-        sheet.add_subview(btn_all)
-
-        btn_none = ui.Button(title="None")
-        btn_none.frame = (100, btn_y, btn_w, btn_h)
-        btn_none.background_color = "#333333"
-        btn_none.tint_color = "white"
-        btn_none.corner_radius = 4
-        btn_none.action = lambda s: toggle_all(False)
-        sheet.add_subview(btn_none)
-
-        # TableView
-        tv_y = 80
-        tv_h = sheet.height - tv_y - 60 # space for bottom bar
-        tv = ui.TableView()
-        tv.frame = (0, tv_y, sheet.width, tv_h)
-        tv.flex = "WH"
-        tv.background_color = "#111111"
-        tv.separator_color = "#333333"
-        tv.allows_multiple_selection = False # We handle selection manually
-
-        class PrescanDS(object):
-            def tableview_number_of_rows(self, tv, section):
-                return len(flat_items)
-
-            def tableview_cell_for_row(self, tv, section, row):
-                item = flat_items[row]
-                cell = ui.TableViewCell()
-                cell.text_label.text = item["display"]
-                cell.text_label.font = ("<Mono>", 12)
-                cell.background_color = "#111111"
-                cell.text_label.text_color = "white" if item["type"] == "file" else "#88ccff"
-
-                if item["selected"]:
-                    cell.accessory_type = "checkmark"
-                else:
-                    cell.accessory_type = "none"
-                return cell
-
-            def tableview_did_select(self, tv, section, row):
-                # Toggle logic
-                item = flat_items[row]
-                new_state = not item["selected"]
-                
-                # Handle ALL state transition
-                if selection_state['mode'] == 'all' and not new_state:
-                    # Deselecting from ALL state - switch to partial selection mode
-                    selection_state['mode'] = 'partial'
-                
-                self._set_selected_recursive(item, new_state)
-                
-                # Update selection mode after change
-                if all(i["selected"] for i in flat_items):
-                    selection_state['mode'] = 'all'
-                elif not any(i["selected"] for i in flat_items):
-                    selection_state['mode'] = 'none'
-                else:
-                    selection_state['mode'] = 'partial'
-                
-                tv.reload_data()
-
-            def _set_selected_recursive(self, item, state):
-                item["selected"] = state
-                # If dir, find children in flat list and toggle
-                if item["type"] == "dir":
-                    # Naive: scan following items with depth > item.depth
-                    # Since it is a flat list from traversal, children are immediately following.
-                    idx = flat_items.index(item)
-                    for i in range(idx + 1, len(flat_items)):
-                        child = flat_items[i]
-                        if child["depth"] <= item["depth"]:
-                            break
-                        child["selected"] = state
-
-        ds = PrescanDS()
-        tv.data_source = ds
-        tv.delegate = ds
-        sheet.add_subview(tv)
-
-        # Bottom Bar: Remove / Cancel / Replace / Append
-        bar_y = sheet.height - 50
-        
-        # Shared pool update logic
-        def _pool_update(mode):
-            """
-            Update the prescan selection pool.
-            mode: 'replace', 'append', or 'remove'
-            """
-            # Create a map for quick lookup of selection status by path
-            # Note: We rely on the UI state (flat_items) where possible.
-            selection_map = {item["path"]: item["selected"] for item in flat_items}
-
-            # FIX 1: Materialize raw paths correctly (DFS).
-            # If a directory is selected, ALL its descendants must be in raw_paths.
-            # We cannot rely solely on flat_items["selected"] for files if the user only clicked the folder.
-            
-            materialized_raw = []
-            compressed_paths = []
-
-            def collect_materialized(node):
-                path = node["path"]
-                # Check selection state from map (populated by UI toggles)
-                is_selected = selection_map.get(path, False)
-
-                if is_selected:
-                    if node["type"] == "file":
-                        materialized_raw.append(path)
-                        compressed_paths.append(path)
-                    else:
-                        # Directory is selected -> fully selected
-                        compressed_paths.append(path)
-                        # Materialize all descendants for raw truth
-                        collect_all_descendants(node)
-                else:
-                    # Not selected -> descend
-                    if node["type"] == "dir" and node.get("children"):
-                        for c in node["children"]:
-                            collect_materialized(c)
-
-            def collect_all_descendants(node):
-                if node["type"] == "file":
-                    materialized_raw.append(node["path"])
-                elif node.get("children"):
-                    for c in node["children"]:
-                        collect_all_descendants(c)
-
-            collect_materialized(root_node)
-
-            # Normalize and deduplicate
-            raw_paths = sorted(list(set(normalize_path(p) for p in materialized_raw)))
-            compressed_paths = sorted(list(set(normalize_path(p) for p in compressed_paths)))
-            
-            # Handle different modes
-            if mode == 'remove':
-                # Remove from pool
-                if root_name in self.saved_prescan_selections:
-                    del self.saved_prescan_selections[root_name]
-                self.save_last_state()
-                if console:
-                    console.hud_alert(f"Removed selection pool for {root_name}", "success", 1.5)
-                reset_guard()
-                sheet.close()
-                return
-            
-            # Get current selection mode
-            current_mode = selection_state['mode']
-            
-            # Check if we have an existing selection for this repo
-            existing = self.saved_prescan_selections.get(root_name)
-            
-            if mode == 'replace':
-                # Replace mode: overwrite existing selection
-                if current_mode == 'all':
-                    # ALL selected
-                    self.saved_prescan_selections[root_name] = {"raw": None, "compressed": None}
-                elif current_mode == 'none':
-                    # Nothing selected - remove from pool
-                    if root_name in self.saved_prescan_selections:
-                        del self.saved_prescan_selections[root_name]
-                else:
-                    # Partial selection - store both raw and compressed
-                    if compressed_paths or raw_paths:
-                        self.saved_prescan_selections[root_name] = {
-                            "raw": raw_paths if raw_paths else None,
-                            "compressed": compressed_paths if compressed_paths else None
-                        }
-                    else:
-                        # Empty selection - remove from pool
-                        if root_name in self.saved_prescan_selections:
-                            del self.saved_prescan_selections[root_name]
-                
-                self.save_last_state()
-                if console:
-                    console.hud_alert(f"Replaced selection pool for {root_name}", "success", 1.5)
-                
-            elif mode == 'append':
-                # Append mode: union with existing selection
-                if current_mode == 'none':
-                    # Nothing selected in current view - no-op with feedback
-                    if console:
-                        console.hud_alert("No changes: no items selected in append mode", "error", 2.0)
-                    return # Don't close dialog
-                
-                if current_mode == 'all':
-                    # ALL selected - ALL overrides everything
-                    self.saved_prescan_selections[root_name] = {"raw": None, "compressed": None}
-                else:
-                    # Partial selection - union raw, then RE-COMPRESS
-                    merged_raw = None
-
-                    if existing and isinstance(existing, dict):
-                        existing_raw = existing.get("raw")
-                        if existing_raw is None:
-                            # Existing was ALL. Union(ALL, Partial) = ALL
-                            self.saved_prescan_selections[root_name] = {"raw": None, "compressed": None}
-                            self.save_last_state()
-                            if console: console.hud_alert(f"Appended to selection pool for {root_name}", "success", 1.5)
-                            reset_guard()
-                            sheet.close()
-                            return
-                        else:
-                            # Union of existing and new raw paths
-                            merged_raw = set(existing_raw)
-                            if raw_paths:
-                                merged_raw.update(raw_paths)
-                    else:
-                        # No existing -> just new raw
-                        merged_raw = set(raw_paths) if raw_paths else None
-
-                    # If we have a merged raw set, re-compress using the tree (Iterative DFS)
-                    if merged_raw and len(merged_raw) > 0:
-                        new_compressed = []
-
-                        # Build a map for O(1) lookup, ensure normalization
-                        raw_set = set(normalize_path(p) for p in merged_raw)
-
-                        # Phase 1: Determine selection status of all nodes (Post-order simulation)
-                        # We need to know if a dir is fully selected. Since flat_items is flat,
-                        # we can't easily do post-order without recursion.
-                        # BUT: flat_items was built via DFS. We can iterate backwards?
-                        # No, simpler: Build a tree-like structure or map from the flat items?
-                        # Actually, root_node is available and it IS a tree.
-
-                        # Iterative Post-Order to mark 'fully_selected'
-                        # We decorate the nodes temporarily or use a map ID->Status
-
-                        node_status = {} # path -> bool (fully selected)
-
-                        # Iterative Post-Order Traversal using 2 stacks
-                        stack1 = [root_node]
-                        stack2 = []
-                        while stack1:
-                            node = stack1.pop()
-                            stack2.append(node)
-                            if node.get("children"):
-                                for c in node["children"]:
-                                    stack1.append(c)
-
-                        # Process stack2 (children before parents)
-                        while stack2:
-                            node = stack2.pop()
-                            path = normalize_path(node["path"])
-
-                            if node["type"] == "file":
-                                node_status[path] = path in raw_set
-                            else: # dir
-                                children = node.get("children", [])
-                                if not children:
-                                    node_status[path] = False # Empty dir not selected
-                                else:
-                                    # All children must be fully selected
-                                    all_selected = True
-                                    for c in children:
-                                        c_path = normalize_path(c["path"])
-                                        if not node_status.get(c_path, False):
-                                            all_selected = False
-                                            break
-                                    node_status[path] = all_selected
-
-                        # Phase 2: Collect compressed paths (Pre-order)
-                        # If a node is fully selected, add it and skip children. Else descend.
-                        stack = [root_node]
-                        while stack:
-                            node = stack.pop()
-                            path = normalize_path(node["path"])
-
-                            if node_status.get(path, False):
-                                # Fully selected (Dir or File)
-                                new_compressed.append(path)
-                                # Do NOT push children
-                            else:
-                                # Not fully selected. If dir, push children to check them.
-                                # Push in reverse order to maintain order when popping
-                                if node.get("children"):
-                                    for i in range(len(node["children"]) - 1, -1, -1):
-                                        stack.append(node["children"][i])
-                                elif node["type"] == "file":
-                                    # File not selected? Then don't include.
-                                    # (Should be covered by node_status check above, but logic:
-                                    # if file is false, we do nothing)
-                                    pass
-
-                        self.saved_prescan_selections[root_name] = {
-                            "raw": sorted(list(raw_set)),
-                            "compressed": [normalize_path(p) for p in new_compressed]
-                        }
-                    else:
-                        # Empty result -> remove
-                        if root_name in self.saved_prescan_selections:
-                            del self.saved_prescan_selections[root_name]
-                
-                self.save_last_state()
-                if console:
-                    console.hud_alert(f"Appended to selection pool for {root_name}", "success", 1.5)
-            
-            reset_guard()
-            sheet.close()
-            # No auto-merge!
-        
-        # Remove button (left side)
-        btn_remove = ui.Button(title="Remove")
-        btn_remove.frame = (10, bar_y, 80, 40)
-        btn_remove.flex = "RT" # Right margin flex, Top margin flex
-        btn_remove.background_color = "#ff3b30"
-        btn_remove.tint_color = "white"
-        btn_remove.corner_radius = 6
-        btn_remove.action = lambda s: _pool_update('remove')
-        sheet.add_subview(btn_remove)
-        
-        # Cancel button (right side)
-        btn_cancel = ui.Button(title="Cancel")
-        btn_cancel.frame = (sheet.width - 310, bar_y, 70, 40)
-        btn_cancel.flex = "LT" # Left margin flex, Top margin flex
-        btn_cancel.background_color = "#444444"
-        btn_cancel.tint_color = "white"
-        btn_cancel.corner_radius = 6
-        btn_cancel.action = lambda s: (reset_guard(), sheet.close())
-        sheet.add_subview(btn_cancel)
-        
-        # Replace button (right side)
-        # "Store to Pool (Replace)" - abbreviated for mobile UI if needed, but clarity is prioritized
-        btn_replace = ui.Button(title="Store (Replace)")
-        btn_replace.frame = (sheet.width - 250, bar_y, 120, 40)
-        btn_replace.flex = "LT" # Left margin flex, Top margin flex
-        btn_replace.background_color = "#007aff"
-        btn_replace.tint_color = "white"
-        btn_replace.corner_radius = 6
-        btn_replace.action = lambda s: _pool_update('replace')
-        sheet.add_subview(btn_replace)
-        
-        # Append button (right side)
-        # "Store to Pool (Append)"
-        btn_append = ui.Button(title="Store (Append)")
-        btn_append.frame = (sheet.width - 120, bar_y, 110, 40)
-        btn_append.flex = "LT" # Left margin flex, Top margin flex
-        btn_append.background_color = "#34c759"
-        btn_append.tint_color = "white"
-        btn_append.corner_radius = 6
-        btn_append.action = lambda s: _pool_update('append')
-        sheet.add_subview(btn_append)
-
-        sheet.present("sheet")
-
-
-    def show_pool_viewer(self, sender):
-        """Shows the current Selection Pool content."""
-        pool = self.saved_prescan_selections
-
-        sheet = ui.View()
-        sheet.name = "Selection Pool"
-        sheet.background_color = "#111111"
-        sheet.frame = (0, 0, 500, 600)
-
-        if not pool:
-            lbl = ui.Label(frame=(0, 0, 500, 600))
-            lbl.text = "Pool is empty."
-            lbl.alignment = ui.ALIGN_CENTER
-            lbl.text_color = "gray"
-            sheet.add_subview(lbl)
-        else:
-            tv = ui.TableView()
-            tv.frame = (0, 0, 500, 540)
-            tv.flex = "WH"
-            tv.background_color = "#111111"
-            tv.separator_color = "#333333"
-
-            # Helper to format info
-            def format_pool_info(data):
-                if not data or not isinstance(data, dict):
-                    return "Invalid state"
-
-                # Check for ALL state (both None)
-                raw = data.get("raw")
-                compressed = data.get("compressed")
-
-                if raw is None and compressed is None:
-                    return "ALL"
-
-                # Partial state
-                raw_count = len(raw) if raw else 0
-                compressed_count = len(compressed) if compressed else 0
-                return f"Partial: {raw_count} files / {compressed_count} rules"
-
-            # Convert pool to list
-            items = []
-            for repo, data in pool.items():
-                info = format_pool_info(data)
-                items.append({"repo": repo, "info": info})
-
-            items.sort(key=lambda x: x["repo"])
-
-            class PoolDS(object):
-                def __init__(self, parent_ui):
-                    self.parent = parent_ui
-
-                def tableview_number_of_rows(self, tv, section):
-                    return len(items)
-
-                def tableview_cell_for_row(self, tv, section, row):
-                    item = items[row]
-                    cell = ui.TableViewCell('value1')
-                    cell.text_label.text = item["repo"]
-                    cell.detail_text_label.text = item["info"]
-                    cell.text_label.text_color = "white"
-                    cell.detail_text_label.text_color = "#888888"
-                    cell.background_color = "#111111"
-                    cell.accessory_type = 'detail_button'
-                    return cell
-
-                def tableview_accessory_button_tapped(self, tv, section, row):
-                    self.show_inspector(row)
-
-                def tableview_did_select(self, tv, section, row):
-                    self.show_inspector(row)
-
-                def show_inspector(self, row):
-                    item = items[row]
-                    repo = item["repo"]
-                    entry = pool.get(repo)
-
-                    if not entry or not isinstance(entry, dict):
-                        return
-
-                    raw = entry.get("raw")
-                    compressed = entry.get("compressed")
-
-                    msg = f"Repo: {repo}\n\n"
-
-                    if raw is None and compressed is None:
-                        msg += "State: ALL files included."
-                    else:
-                        r_count = len(raw) if raw else 0
-                        c_count = len(compressed) if compressed else 0
-                        msg += f"State: Partial\nFiles: {r_count}\nRules: {c_count}\n\n"
-
-                        if compressed:
-                            msg += "Rules (Compressed):\n"
-                            # Limit display
-                            display_rules = compressed[:15]
-                            for r in display_rules:
-                                msg += f"- {r}\n"
-                            if len(compressed) > 15:
-                                msg += f"... and {len(compressed)-15} more"
-
-                    if console:
-                        console.alert("Pool Details", msg, "OK", hide_cancel_button=True)
-                    elif ui:
-                        ui.alert("Pool Details", msg, "OK")
-
-                def tableview_can_edit(self, tv, section, row):
-                    return True
-
-                def tableview_delete(self, tv, section, row):
-                    repo = items[row]["repo"]
-                    if repo in pool:
-                        del pool[repo]
-                        # Persist immediately using parent reference
-                        self.parent.save_last_state()
-                        self.parent._update_repo_info()
-
-                    items.pop(row)
-                    tv.delete_rows([row])
-
-            ds = PoolDS(self)
-            tv.data_source = ds
-            tv.delegate = ds
-            sheet.add_subview(tv)
-
-            # Bottom Bar
-            bar = ui.View(frame=(0, 540, 500, 60))
-            bar.flex = "WT"
-            bar.background_color = "#222222"
-
-            btn_clear = ui.Button(title="Clear Pool")
-            btn_clear.frame = (10, 10, 100, 40)
-            btn_clear.background_color = "#ff3b30"
-            btn_clear.tint_color = "white"
-            btn_clear.corner_radius = 6
-            def clear_action(sender):
-                pool.clear()
-                self.save_last_state()
-                self._update_repo_info()
-                sheet.close()
-                if console: console.hud_alert("Pool cleared")
-            btn_clear.action = clear_action
-            bar.add_subview(btn_clear)
-
-            sheet.add_subview(bar)
-
-        sheet.present("sheet")
-
-    def _run_merge_form_reset_after_success_safe(self) -> None:
-        """Safely reset form after success, with error handling and logging."""
-        try:
-            self.reset_merge_form_to_defaults_after_success()
-        except Exception as exc:
-            print(
-                f"[RepoGround] Warning: failed to reset merge form after success: {exc}",
-                file=sys.stderr,
-            )
-
-    def schedule_merge_form_reset_after_success(self) -> None:
-        """Schedule form reset on main thread (UI thread) if possible."""
-        if ui is not None and hasattr(ui, "delay"):
-            ui.delay(self._run_merge_form_reset_after_success_safe, 0.0)
-        else:
-            self._run_merge_form_reset_after_success_safe()
-
-    def reset_merge_form_to_defaults_after_success(self) -> None:
-        """Reset only merge-related transient UI state after a successful merge run."""
-        pool = getattr(self, "saved_prescan_selections", None)
-        if pool is None:
-            self.saved_prescan_selections = {}
-        else:
-            pool.clear()
-
-        try:
-            if getattr(self, "tv", None) is not None:
-                self.tv.selected_rows = []
-                if hasattr(self.tv, "reload_data"):
-                    self.tv.reload_data()
-        except Exception as exc:
-            print(
-                f"[RepoGround] Warning: failed to reset table view selection: {exc}",
-                file=sys.stderr,
-            )
-
-        if getattr(self, "ext_field", None) is not None:
-            self.ext_field.text = ""
-        if getattr(self, "path_field", None) is not None:
-            self.path_field.text = ""
-
-        if getattr(self, "max_field", None) is not None:
-            self.max_field.text = "" if DEFAULT_MAX_FILE_BYTES <= 0 else str(DEFAULT_MAX_FILE_BYTES)
-
-        if getattr(self, "split_field", None) is not None:
-            split_text = ""
-            if DEFAULT_SPLIT_SIZE and str(DEFAULT_SPLIT_SIZE).strip() not in ("", "0"):
-                raw = str(DEFAULT_SPLIT_SIZE).strip()
-                if raw.isdigit():
-                    split_text = raw
-                else:
-                    try:
-                        mb = int(round(parse_human_size(raw) / (1024 * 1024)))
-                        split_text = str(mb) if mb > 0 else ""
-                    except Exception:
-                        split_text = raw
-            self.split_field.text = split_text
-
-        if getattr(self, "seg_detail", None) is not None:
-            try:
-                self.seg_detail.selected_index = self.seg_detail.segments.index(DEFAULT_LEVEL)
-            except Exception:
-                self.seg_detail.selected_index = 0
-            if hasattr(self, "on_profile_changed"):
-                self.on_profile_changed(None)
-
-        if getattr(self, "seg_mode", None) is not None:
-            self.seg_mode.selected_index = 1 if DEFAULT_MODE == "pro-repo" else 0
-
-        if getattr(self, "seg_meta", None) is not None:
-            try:
-                self.seg_meta.selected_index = self.seg_meta.segments.index(DEFAULT_META_DENSITY)
-            except Exception:
-                self.seg_meta.selected_index = 0
-
-        if getattr(self, "plan_only_switch", None) is not None:
-            self.plan_only_switch.value = False
-        if getattr(self, "code_only_switch", None) is not None:
-            self.code_only_switch.value = False
-        if getattr(self, "pre_pull_switch", None) is not None:
-            self.pre_pull_switch.value = True
-
-        extras_defaults, _ = ExtrasConfig.from_csv(DEFAULT_EXTRAS)
-        self.extras_config = extras_defaults
-
-        self._update_repo_info()
-        self.save_last_state()
-
-    def run_merge(self, sender) -> None:
-        """
-        UI-Handler: niemals schwere Arbeit im Main-Thread ausführen,
-        sonst wirkt Pythonista "eingefroren" – besonders bei Multi-Repo.
-        """
-        if getattr(self, "_prescan_active", False):
-            if console:
-                console.hud_alert("Prescan active - merge blocked", "error")
-            return
-
-        # Snapshot UI state on main thread to avoid thread-safety issues in background
-        self._pending_plan_only = self.plan_only_switch.value
-        # Use getattr for code_only just in case (legacy robustness)
-        self._pending_code_only = bool(getattr(self, "code_only_switch", None) and self.code_only_switch.value)
-        self._pending_pre_pull = resolve_pre_pull_switch_value(getattr(self, "pre_pull_switch", None))
-
-        try:
-            import ui as _ui
-            in_bg = getattr(_ui, "in_background", None)
-        except Exception:
-            in_bg = None
-
-        if in_bg:
-            in_bg(self._run_merge_safe)()
-        else:
-            # Fallback: wenigstens nicht crashen – aber UI kann dann weiterhin blocken.
-            self._run_merge_safe()
-
-    def _run_merge_safe(self) -> None:
-        try:
-            # Aktuellen Zustand merken
-            self.save_last_state()
-            self._run_merge_inner()
-        except Exception as e:
-            traceback.print_exc()
-
-            # Use specific messaging for validation errors if possible
-            if "ValidationException" in type(e).__name__ or "Structure Violation" in str(e):
-                msg = f"Validation Error: {e}"
-            else:
-                msg = f"Error: {e}"
-
-            if console:
-                console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-            else:
-                print(msg, file=sys.stderr)
-        finally:
-            # Cleanup snapshotted state to prevent stale values in future runs
-            if hasattr(self, "_pending_plan_only"):
-                del self._pending_plan_only
-            if hasattr(self, "_pending_code_only"):
-                del self._pending_code_only
-            if hasattr(self, "_pending_pre_pull"):
-                del self._pending_pre_pull
-
-    def _run_merge_inner(self) -> None:
-        # 1. Determine Selection Strategy (Explicit vs Pool vs All)
-        tv = self.tv
-        rows = tv.selected_rows or []
-
-        selected_repos: List[str] = []
-        selection_source = "default(all)"
-
-        # Build normalized maps once (avoid N^2 drift pain)
-        repo_norm_map = {}
-        for r in self.repos:
-            norm = normalize_repo_id(r)
-            if norm in repo_norm_map:
-                # Collision detected
-                msg = f"Repo collision: '{r}' and '{repo_norm_map[norm]}' normalize to same ID '{norm}'."
-                print(f"[RepoGround] WARNING: {msg}", file=sys.stderr)
-            else:
-                repo_norm_map[norm] = r
-
-        pool_raw = getattr(self, "saved_prescan_selections", None) or {}
-        pool_norm = {}
-        if isinstance(pool_raw, dict):
-            for k, v in pool_raw.items():
-                n_key = normalize_repo_id(k)
-                if n_key: # Ignore empty keys
-                    if n_key in pool_norm:
-                        print(f"[RepoGround] WARNING: Pool key collision for '{n_key}' (overwriting).", file=sys.stderr)
-                    pool_norm[n_key] = v
-
-        # Pool repos that actually exist in current hub
-        pool_active_repos = sorted(list({repo_norm_map[nk] for nk in pool_norm.keys() if nk in repo_norm_map}))
-
-        if rows:
-            # Explicit UI selection
-            selection_source = "explicit"
-            for section, row in rows:
-                if 0 <= row < len(self.repos):
-                    selected_repos.append(self.repos[row])
-        elif pool_active_repos:
-            # Implicit Pool selection
-            selection_source = "pool"
-            selected_repos = pool_active_repos
-        else:
-            # Default All
-            selection_source = "default(all)"
-            selected_repos = list(self.repos)
-
-        # Deduplicate and Sort
-        selected_repos = sorted(list(set(selected_repos)))
-
-        if not selected_repos:
-            if console:
-                console.alert("RepoGround", "No repos selected or found in pool.", "OK", hide_cancel_button=True)
-            return
-
-        ext_text = (self.ext_field.text or "").strip()
-        extensions = _normalize_ext_list(ext_text)
-
-        path_contains = (self.path_field.text or "").strip() or None
-
-        detail_idx = self.seg_detail.selected_index
-        detail = ["overview", "summary", "dev", "max"][detail_idx]
-
-        mode_idx = self.seg_mode.selected_index
-        mode = ["gesamt", "pro-repo"][mode_idx]
-
-        meta_idx = self.seg_meta.selected_index
-        meta_density = self.seg_meta.segments[meta_idx]
-
-        max_bytes = self._parse_max_bytes()
-        split_size = self._parse_split_size()
-
-        # Use snapshotted values from main thread if available (thread-safe), else fallback
-        if hasattr(self, "_pending_plan_only"):
-            plan_only = self._pending_plan_only
-        else:
-            plan_switch = getattr(self, "plan_only_switch", None)
-            plan_only = bool(plan_switch and plan_switch.value)
-
-        if hasattr(self, "_pending_code_only"):
-            code_only = self._pending_code_only
-        else:
-            code_switch = getattr(self, "code_only_switch", None)
-            code_only = bool(code_switch and code_switch.value)
-
-        if hasattr(self, "_pending_pre_pull"):
-            pre_pull = self._pending_pre_pull
-        else:
-            pre_pull_switch = getattr(self, "pre_pull_switch", None)
-            pre_pull = bool(pre_pull_switch is None or pre_pull_switch.value)
-
-        # Mutual exclusion: plan_only wins to avoid ambiguous semantics.
-        if plan_only and code_only:
-            code_only = False
-
-        # Helper to get include_paths from pool with normalized O(1) lookup
-        def get_pool_include_paths(repo_name):
-            entry = pool_norm.get(normalize_repo_id(repo_name))
-            return resolve_pool_include_paths(entry)
-
-        # Calculate Stats for UX
-        total_paths = 0
-        repos_with_filters = 0
-        for name in selected_repos:
-            paths = get_pool_include_paths(name)
-            if paths is not None:
-                if isinstance(paths, list):
-                    total_paths += len(paths)
-                repos_with_filters += 1
-
-        # HUD / Log Feedback
-        pool_keys_total = len(pool_norm) if isinstance(pool_norm, dict) else 0
-        pool_keys_matched = len(pool_active_repos)
-
-        msg = f"Selection: {selection_source.upper()} ({len(selected_repos)} repos)"
-        if selection_source == "pool":
-            msg += f" / pool matched {pool_keys_matched}/{pool_keys_total}"
-
-        if repos_with_filters > 0:
-            if selection_source == "pool":
-                msg += f" / {total_paths} paths"
-            else:
-                msg += f" / {total_paths} paths from pool"
-
-        if console:
-            console.hud_alert(msg, "info", 1.5)
-        else:
-            print(f"[RepoGround] {msg}")
-
-        # Check for restrictive pool entries to decide on execution strategy
-        has_restrictive = False
-        for name in selected_repos:
-            paths = get_pool_include_paths(name)
-            if paths is not None and isinstance(paths, list) and len(paths) > 0:
-                has_restrictive = True
-                break
-
-        merges_dir = get_merges_dir(self.hub)
-        all_out_paths = []
-
-        # Pre-pull (bounded repo-sync mutation), two-phase across ALL selected repos
-        # BEFORE any scan, so no repo is fast-forwarded when another repo hard-fails during the plan phase.
-        # plan_only never mutates local repos, so it forces pre-pull off; Pythonista/iOS
-        # forces it off too (no git subprocesses) — see resolve_effective_pre_pull.
-        def _pre_pull_hud(message):
-            try:
-                if console:
-                    console.hud_alert(message, "info", 2.0)
-            except Exception:
-                # Best-effort UI notification only; HUD errors must not interrupt the merge flow.
-                pass
-
-        effective_pre_pull = resolve_effective_pre_pull(
-            pre_pull, plan_only, log=print, notify=_pre_pull_hud
-        )
-        if effective_pre_pull:
-            pre_pull_sources = [self.hub / name for name in selected_repos if (self.hub / name).is_dir()]
-            try:
-                run_pre_pull_two_phase(pre_pull_sources, log=print)
-            except Exception as e:
-                if console:
-                    try:
-                        console.hud_alert(f"Pre-pull failed: {e}", "error", 2.0)
-                    except Exception:
-                        # Best-effort UI notification only; ignore HUD failures and continue
-                        # with stderr logging + re-raising the original pre-pull exception.
-                        pass
-                print(f"Pre-pull failed: {e}", file=sys.stderr)
-                raise
-        elif pre_pull and plan_only:
-            print("Pre-pull skipped because plan_only=True.")
-
-        # Execution Strategy
-        # If restrictive pool entries exist, we must split execution per repo to ensure
-        # correct include_paths are respected and not mixed with global filters.
-        # This matches WebUI "force pro-repo" logic.
-
-        execution_list = []
-        if has_restrictive:
-            # Force sequential processing per repo
-            for name in selected_repos:
-                execution_list.append([name])
-            if console:
-                console.hud_alert("Pool active: Splitting jobs per repo")
-        else:
-            # Standard batch processing (all selected together)
-            execution_list.append(selected_repos)
-
-        for b_idx, batch_repos in enumerate(execution_list, start=1):
-            summaries = []
-
-            # Scan phase for this batch
-            for i, name in enumerate(batch_repos, start=1):
-                root = self.hub / name
-                if not root.is_dir():
-                    continue
-
-                # Pre-pull already handled once (two-phase) before this loop.
-
-                # Feedback
-                if console:
-                    try:
-                        console.hud_alert(f"Scanning {name} ({i}/{len(batch_repos)})", duration=0.6)
-                    except Exception:
-                        pass
-                try:
-                    import ui as _ui
-                    _ui.delay(lambda: None, 0.0)
-                except Exception:
-                    pass
-
-                # Resolve paths
-                use_include_paths = get_pool_include_paths(name)
-
-                # If we are in a restrictive split-job scenario, and this repo has NO pool entry,
-                # it uses global defaults (None). If it HAS an entry, it uses that.
-                # scan_repo handles include_paths=None as "scan all".
-
-                summary = scan_repo(root, extensions or None, path_contains, max_bytes, include_paths=use_include_paths, calculate_md5=True, include_hidden=True)
-                summaries.append(summary)
-
-            if not summaries:
-                continue
-
-            # Delta Logic Injection (Batch-aware)
-            delta_meta = None
-            if self.extras_config.delta_reports and len(summaries) == 1:
-                repo_name = summaries[0]["name"]
-                try:
-                    mod = _load_repoground_extractor_module()
-                    if mod and hasattr(mod, "find_latest_diff_for_repo") and hasattr(mod, "extract_delta_meta_from_diff_file"):
-                        diff_path = mod.find_latest_diff_for_repo(merges_dir, repo_name)
-                        if diff_path:
-                            delta_meta = mod.extract_delta_meta_from_diff_file(diff_path)
-                except Exception as e:
-                    print(f"[RepoGround] Warning: Could not extract delta metadata: {e}", file=sys.stderr)
-
-            # Generate Report for this batch
-            # If splitting jobs, we likely want "pro-repo" mode for the output to be named correctly,
-            # or if it was "gesamt" but forced split, we get one file per repo anyway.
-            # We preserve the user's requested mode unless we forced a split, in which case effectively it behaves like pro-repo
-            # but write_reports_v2 needs to know how to name it.
-            # If has_restrictive is True, we are iterating 1-by-1. calling write_reports_v2 with 1 summary.
-            # Mode "gesamt" with 1 summary produces 1 file. Mode "pro-repo" with 1 summary produces 1 file.
-            # So passing the user's `mode` is fine.
-
-            # However, if we forced split, we should probably ensure `path_filter` passed to write_reports
-            # doesn't confuse the header if we are using specific include_paths.
-            # But `write_reports_v2` uses `path_filter` just for metadata display usually.
-            # We pass it through.
-
-            artifacts = write_reports_v2(
-                merges_dir,
-                self.hub,
-                summaries,
-                detail,
-                mode,
-                max_bytes,
-                plan_only,
-                code_only,
-                split_size,
-                debug=False,
-                path_filter=path_contains,
-                ext_filter=extensions or None,
-                extras=self.extras_config,
-                delta_meta=delta_meta,
-                meta_density=meta_density,
-                generator_info={"name": "repoground", "platform": "ios"},
-            )
-
-            all_out_paths.extend(artifacts.get_all_paths())
-
-            # Force close intermediate files
-            force_close_files(artifacts.get_all_paths())
-
-        if not all_out_paths:
-            if console:
-                console.alert("RepoGround", "No report generated.", "OK", hide_cancel_button=True)
-            else:
-                print("No report generated.")
-            return
-
-        # Generate Bundle Index if multiple artifacts (Post-Step)
-        if len(all_out_paths) > 1:
-            try:
-                now_ts = datetime.datetime.now(datetime.timezone.utc).strftime("%Y-%m-%dT%H%M%SZ")
-                bundle_path = merges_dir / f"bundle-merge_{now_ts}.md"
-
-                # Identify restrictive repos for metadata
-                restrictive_repos = []
-                for name in selected_repos:
-                    paths = get_pool_include_paths(name)
-                    if paths is not None and isinstance(paths, list) and len(paths) > 0:
-                        restrictive_repos.append(name)
-
-                # Dynamic split_mode based on actual execution logic
-                is_split_mode = has_restrictive
-
-                lines = [
-                    "---",
-                    "pool_status:",
-                    f"  restrictive_repos: {json.dumps(restrictive_repos)}",
-                    f"  split_mode: {str(is_split_mode).lower()}",
-                    "---",
-                    "# Bundle Merge Report",
-                    f"- Generated: {now_ts}",
-                    f"- Parts: {len(all_out_paths)}",
-                    "",
-                    "## Index",
-                ]
-
-                for p in all_out_paths:
-                    lines.append(f"- [{p.name}]({p.name})")
-
-                bundle_path.write_text("\n".join(lines), encoding="utf-8")
-                # Prepend to be the primary result
-                all_out_paths.insert(0, bundle_path)
-            except Exception as e:
-                print(f"Failed to create bundle index: {e}", file=sys.stderr)
-
-        # Summary Feedback
-        count = len(all_out_paths)
-        primary = _pick_primary_artifact(all_out_paths)
-
-        # Enforce bundle as primary if present
-        for p in all_out_paths:
-            if "bundle-merge" in p.name:
-                primary = p
-                break
-
-        if count == 1 and primary:
-            msg = f"Merge generated: {primary.name}"
-        elif primary and count > 1:
-            msg = f"Merge generated: {primary.name} (+{count-1} parts)"
-        else:
-            msg = f"Generated {count} artifacts"
-
-        if console:
-            try:
-                console.hud_alert(msg, "success", 1.5)
-            except Exception as e:
-                sys.stderr.write(f"Warning: Failed to show HUD alert (falling back to alert): {e}\n")
-                console.alert("RepoGround", msg, "OK", hide_cancel_button=True)
-        else:
-            print(f"RepoGround: {msg}")
-            for p in all_out_paths:
-                print(f"  - {p.name}")
-
-        if all_out_paths:
-            self.schedule_merge_form_reset_after_success()
 
 
 # --- CLI Mode ---
+
+
+
+def _wire_pythonista_mixin_globals() -> None:
+    """Bind build-module symbols into extracted MergerUI mixins.
+
+    Mixins load during build.py import. After this module finished defining
+    helpers/constants (and optional Pythonista ui), copy those names into the
+    mixin modules so method bodies keep their historical free-variable lookup.
+    """
+    import sys
+
+    package = __package__ or ""
+    module_names = (
+        f"{package}.merger_ui_prescan" if package else "merger_ui_prescan",
+        f"{package}.merger_ui_merge_run" if package else "merger_ui_merge_run",
+        f"{package}.merger_ui_browser" if package else "merger_ui_browser",
+    )
+    skip = {
+        "__name__",
+        "__file__",
+        "__package__",
+        "__cached__",
+        "__builtins__",
+        "__spec__",
+        "__loader__",
+        "__doc__",
+        "__all__",
+        "MergerUI",
+        "MergerUIPrescanMixin",
+        "MergerUIBrowserMixin",
+        "MergerUIMergeRunMixin",
+        "_wire_pythonista_mixin_globals",
+    }
+    payload = {
+        name: value
+        for name, value in globals().items()
+        if name not in skip and not name.startswith("__")
+    }
+    for module_name in module_names:
+        module = sys.modules.get(module_name)
+        if module is None:
+            continue
+        for name, value in payload.items():
+            setattr(module, name, value)
+
+
+_wire_pythonista_mixin_globals()
 
 def _is_headless_requested() -> bool:
     # Headless wenn:
